✨[S12P31A401-93] feat: Implement project list API

+++ spicoBackend/src/main/kotlin/com/ssafy/spico/domain/project/service/ProjectService.kt
@@ -4,5 +4,5 @@ import com.ssafy.spico.domain.project.dto.ProjectViewType
 import com.ssafy.spico.domain.project.model.Project
 
 interface ProjectService {
-    fun getProjects(cursor: Long?, size: Int, type: ProjectViewType) : List<Project>
+    fun getProjects(cursor: Int?, size: Int, type: ProjectViewType) : List<Project>
 }
